done search design, on progress shimmer
