--- core/nav/src/main/java/com/kotlin/learn/core/nav/data/ITransactionNavigationParent.kt
@@ -1,5 +1,0 @@
-package com.kotlin.learn.core.nav.data
-
-interface ITransactionNavigationParent {
-    fun navigateToGreetings() = Unit
-}
