+++ core/nav/src/main/java/com/kotlin/learn/core/nav/di/NavigatorModule.kt
@@ -1,16 +1,11 @@
 package com.kotlin.learn.core.nav.di
 
-import android.app.Activity
-import androidx.fragment.app.FragmentActivity
-import com.kotlin.learn.core.nav.NavControllerBinder
-import com.kotlin.learn.core.nav.navigator.AuthNavigator
-import com.kotlin.learn.core.nav.navigator.MovieNavigator
 import com.kotlin.learn.core.nav.navigator.ParentNavigator
+import com.kotlin.learn.core.nav.navigator.MovieNavigator
+import com.kotlin.learn.core.nav.navigator.MenuNavigator
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
-import dagger.hilt.android.qualifiers.ActivityContext
-import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
 import javax.inject.Singleton
 
@@ -24,14 +19,10 @@ class NavigatorModule {
 
     @Singleton
     @Provides
-    fun provideAuthNavigator(): AuthNavigator = AuthNavigator()
+    fun provideAuthNavigator(): ParentNavigator = ParentNavigator()
 
     @Singleton
     @Provides
-    fun provideParentNavigator(): ParentNavigator = ParentNavigator()
-
-    @Singleton
-    @Provides
-    fun provideBinder(): NavControllerBinder = NavControllerBinder()
+    fun provideParentNavigator(): MenuNavigator = MenuNavigator()
 
 }
