+++ core/nav/src/main/java/com/kotlin/learn/core/nav/navigator/ParentNavigator.kt
@@ -1,19 +1,77 @@
 package com.kotlin.learn.core.nav.navigator
 
-import androidx.fragment.app.FragmentActivity
+import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.NavHostFragment
-import androidx.navigation.fragment.findNavController
-import com.kotlin.learn.core.nav.R
+import com.kotlin.learn.feature.auth.presentation.ui.AuthFragmentDirections
+import com.kotlin.learn.feature.auth.presentation.ui.GreetingsFragmentDirections
+import com.kotlin.learn.feature.auth.presentation.ui.RegisterFragmentDirections
 import com.kotlin.learn.feature.menu.presentation.ui.MenuFragmentDirections
+import com.kotlin.learn.feature.splash.presentation.ui.SplashFragmentDirections
 
 class ParentNavigator {
 
-    private fun initNavHostFragment(activity: FragmentActivity): NavHostFragment {
-        return activity.supportFragmentManager.findFragmentById(R.id.nav_host_fragment_container) as NavHostFragment
+    fun fromSplashToGreetings(fragment: Fragment) {
+        NavHostFragment.findNavController(fragment).navigate(
+            SplashFragmentDirections.actionSplashFragmentToGreetingsFragment()
+        )
     }
 
-    fun fromMenuToGreetings(activity: FragmentActivity) {
-        initNavHostFragment(activity).findNavController().navigate(
+    fun fromSplashToMenu(fragment: Fragment) {
+        NavHostFragment.findNavController(fragment).navigate(
+            SplashFragmentDirections.actionSplashFragmentToMenuFragment()
+        )
+    }
+
+    fun fromGreetingsToRegister(fragment: Fragment) {
+        NavHostFragment.findNavController(fragment).navigate(
+            GreetingsFragmentDirections.actionGreetingsFragmentToRegisterFragment()
+        )
+    }
+
+    fun fromGreetingsToAuth(fragment: Fragment) {
+        NavHostFragment.findNavController(fragment).navigate(
+            GreetingsFragmentDirections.actionGreetingsFragmentToAuthFragment()
+        )
+    }
+
+    fun fromGreetingsToMenu(fragment: Fragment) {
+        NavHostFragment.findNavController(fragment).navigate(
+            GreetingsFragmentDirections.actionGreetingsFragmentToMenuFragment()
+        )
+    }
+
+    fun fromRegisterToGreetings(fragment: Fragment) {
+        NavHostFragment.findNavController(fragment).navigate(
+            RegisterFragmentDirections.actionRegisterFragmentToGreetingsFragment()
+        )
+    }
+
+    fun fromRegisterToAuth(fragment: Fragment) {
+        NavHostFragment.findNavController(fragment).navigate(
+            RegisterFragmentDirections.actionRegisterFragmentToAuthFragment()
+        )
+    }
+
+    fun fromAuthToGreetings(fragment: Fragment) {
+        NavHostFragment.findNavController(fragment).navigate(
+            AuthFragmentDirections.actionAuthFragmentToGreetingsFragment()
+        )
+    }
+
+    fun fromAuthToRegister(fragment: Fragment) {
+        NavHostFragment.findNavController(fragment).navigate(
+            AuthFragmentDirections.actionAuthFragmentToRegisterFragment()
+        )
+    }
+
+    fun fromAuthToMenu(fragment: Fragment) {
+        NavHostFragment.findNavController(fragment).navigate(
+            AuthFragmentDirections.actionAuthFragmentToMenuFragment()
+        )
+    }
+
+    fun fromMenuToGreetings(fragment: Fragment) {
+        NavHostFragment.findNavController(fragment).navigate(
             MenuFragmentDirections.actionMenuFragmentToGreetingsFragment()
         )
     }
