+++ core/nav/src/main/res/navigation/nav_graph_menu.xml
@@ -39,9 +39,9 @@
         tools:layout="@layout/fragment_favourite" />
 
     <fragment
-        android:id="@+id/settingFragment"
-        android:name="com.kotlin.learn.feature.movie.presentation.ui.SettingFragment"
-        tools:layout="@layout/fragment_setting" >
+        android:id="@+id/profileFragment"
+        android:name="com.kotlin.learn.feature.movie.presentation.ui.ProfileFragment"
+        tools:layout="@layout/fragment_profile" >
     </fragment>
 
     <!--    navigation to detail movie-->
