--- feature/menu/src/main/java/com/kotlin/learn/feature/menu/presentation/ui/ProfileFragment.kt
@@ -1,12 +1,0 @@
-package com.kotlin.learn.feature.menu.presentation.ui
-
-import com.kotlin.learn.core.common.base.BaseFragment
-import com.kotlin.learn.feature.menu.databinding.FragmentProfileBinding
-
-class ProfileFragment : BaseFragment<FragmentProfileBinding>(FragmentProfileBinding::inflate) {
-
-    override fun setupView() {
-
-    }
-
-}
