+++ feature/menu/src/main/java/com/kotlin/learn/feature/menu/presentation/ui/SettingFragment.kt
@@ -1,0 +1,12 @@
+package com.kotlin.learn.feature.menu.presentation.ui
+
+import com.kotlin.learn.core.common.base.BaseFragment
+import com.kotlin.learn.feature.movie.databinding.FragmentProfileBinding
+
+class SettingFragment : BaseFragment<FragmentProfileBinding>(FragmentProfileBinding::inflate) {
+
+    override fun setupView() {
+
+    }
+
+}
