+++ feature/menu/src/main/java/com/kotlin/learn/feature/menu/presentation/viewmodel/SettingViewModel.kt
@@ -1,4 +1,4 @@
 package com.kotlin.learn.feature.menu.presentation.viewmodel
 
-class ProfileViewModel {
+class SettingViewModel {
 }
