+++ feature/menu/src/main/res/menu/bottom_nav_menu.xml
@@ -17,7 +17,7 @@
         android:icon="@drawable/ic_favourite"
         android:title="@string/menu_favourite" />
     <item
-        android:id="@+id/settingFragment"
+        android:id="@+id/profileFragment"
         android:icon="@drawable/ic_profile"
         android:title="@string/menu_profile" />
 </menu>
