+++ feature/movie/src/main/res/layout/movie_search_item.xml
@@ -2,43 +2,38 @@
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    android:layout_marginBottom="@dimen/dp_16"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content">
+    android:layout_width="wrap_content"
+    android:layout_height="wrap_content"
+    android:layout_marginRight="@dimen/dp_8"
+    android:layout_marginBottom="@dimen/dp_8">
 
-    <ImageView
-        android:id="@+id/iv_search"
-        android:layout_width="@dimen/dp_140"
-        android:layout_height="@dimen/dp_80"
+    <com.google.android.material.imageview.ShapeableImageView
+        android:id="@+id/thumbnail"
+        android:layout_width="120dp"
+        android:layout_height="180dp"
         android:scaleType="fitXY"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
-        tools:src="@tools:sample/backgrounds/scenic" />
+        app:shapeAppearance="@style/ShapeAppearance.Image.Movie"
+        tools:src="@tools:sample/avatars" />
 
     <TextView
-        android:id="@+id/et_search_title"
-        style="@style/TextAppearance.AppCompat.Small"
+        android:id="@+id/tv_title"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
-        android:layout_marginStart="@dimen/dp_16"
-        android:maxLines="3"
-        android:paddingEnd="@dimen/dp_8"
+        android:layout_marginStart="@dimen/dp_2"
+        android:layout_marginTop="@dimen/dp_4"
+        android:layout_marginEnd="@dimen/dp_2"
+        android:layout_marginBottom="@dimen/dp_4"
+        android:fontFamily="@font/tt_interphases_regular"
+        android:gravity="center"
+        android:maxLines="1"
+        android:textSize="@dimen/sp_12"
         android:textStyle="bold"
-        app:layout_constraintBottom_toBottomOf="@+id/iv_search"
-        app:layout_constraintEnd_toStartOf="@+id/iv_search_icon"
-        app:layout_constraintStart_toEndOf="@+id/iv_search"
-        app:layout_constraintTop_toTopOf="@+id/iv_search"
-        tools:text="Fast and Furious 5" />
-
-    <ImageView
-        android:id="@+id/iv_search_icon"
-        android:layout_width="@dimen/dp_40"
-        android:layout_height="@dimen/dp_40"
-        android:layout_marginEnd="@dimen/dp_8"
-        android:scaleType="fitXY"
-        android:src="@drawable/ic_action_play"
-        app:layout_constraintBottom_toBottomOf="@+id/iv_search"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toTopOf="@+id/iv_search" />
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="@+id/thumbnail"
+        app:layout_constraintStart_toStartOf="@+id/thumbnail"
+        app:layout_constraintTop_toBottomOf="@+id/thumbnail"
+        tools:text="Fast X - Search my movie" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
