+++ feature/movie/src/main/res/layout/see_all_credits_item.xml
@@ -8,8 +8,8 @@
     <androidx.cardview.widget.CardView
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:layout_marginLeft="@dimen/dp_8"
-        android:layout_marginRight="@dimen/dp_8"
+        android:layout_marginLeft="@dimen/dp_4"
+        android:layout_marginRight="@dimen/dp_4"
         android:layout_marginBottom="@dimen/dp_8"
         android:divider="@android:color/transparent"
         android:elevation="@dimen/dp_2"
@@ -27,7 +27,7 @@
             <com.google.android.material.imageview.ShapeableImageView
                 android:id="@+id/iv_thumbnail"
                 android:layout_width="match_parent"
-                android:layout_height="@dimen/dp_300"
+                android:layout_height="180dp"
                 android:scaleType="fitXY"
                 app:layout_constraintBottom_toTopOf="@+id/tv_name"
                 app:layout_constraintStart_toStartOf="parent"
@@ -43,10 +43,11 @@
                 android:layout_marginStart="@dimen/dp_2"
                 android:layout_marginTop="4dp"
                 android:layout_marginEnd="@dimen/dp_2"
+                android:fontFamily="@font/tt_interphases_regular"
                 android:gravity="center"
                 android:lines="1"
                 android:maxLines="1"
-                android:textSize="@dimen/sp_18"
+                android:textSize="@dimen/sp_16"
                 app:layout_constraintBottom_toBottomOf="parent"
                 app:layout_constraintEnd_toEndOf="@+id/iv_thumbnail"
                 app:layout_constraintStart_toStartOf="@+id/iv_thumbnail"
