+++ feature/movie/src/main/res/values/strings.xml
@@ -1,5 +1,6 @@
 <resources>
     <!-- TODO: Remove or change this placeholder text -->
     <string name="hello_blank_fragment">Hello blank fragment</string>
+    <string name="under_development">Under Development</string>
     <string name="banner_home_desc_hint">Nobar Livestream Ashiap man 2022 disini Nobar Livestream Ashiap man 2022 disiniNobar Livestream Ashiap man 2022 disiniNobar Livestream Ashiap man 2022 disini</string>
 </resources>
